Rework file download
