+++ lib/path/parsers/archive.js
@@ -18,8 +18,7 @@ async function parseArchive(token, options) {
   token.children = [{
     inputType: 'path',
     location,
-    temporary: true,
-    filePath: token.filePath
+    filePath: token.fileName
   }]
   token.parsed = true
 }
