+++ lib/url/parsers/archive.js
@@ -1,63 +1,31 @@
 'use strict'
 
-const fs = require('fs')
-const path = require('path')
-const {pipe} = require('mississippi')
-
-const {bytesLimit, computeDigest} = require('../../util/streams')
+const download = require('../download')
 const {isArchive} = require('../../util/archive')
 const {createTempDirectory} = require('../../util/tmpdir')
-
-const {getDigestString} = require('../digest')
-
-async function downloadFile(token, options) {
-  const temp = await createTempDirectory()
-  const filePath = path.join(temp, token.fileName)
-
-  return new Promise((resolve, reject) => {
-    let digest
-    let fileSize
-
-    const onDigest = result => {
-      digest = getDigestString(result.digest, options.digestAlgorithm)
-      fileSize = result.bodyLength
-    }
-
-    pipe(
-      token.response.body,
-      computeDigest(options.digestAlgorithm, onDigest),
-      bytesLimit(100 * 1024 * 1024),
-      fs.createWriteStream(filePath),
-      err => {
-        if (err) return reject(err)
-        resolve({filePath, digest, fileSize})
-      }
-    )
-  })
-}
 
 async function parseArchive(token, options) {
   if (token.parsed) return false
   if (!isArchive(token)) return false
 
   if (!options.extractArchives) {
-    token.response.destroy()
     token.type = 'archive'
     token.parsed = true
     return
   }
 
+  token.temporary = await createTempDirectory()
+
   options.logger.log('archive:download:start', token)
-  const {filePath, digest} = await downloadFile(token, options)
+  const {digest, fileSize} = await download(token, options)
   options.logger.log('archive:download:end', token)
 
   token.digest = digest
+  token.fileSize = fileSize
   token.type = 'archive'
   token.children = [{
     inputType: 'path',
-    location: filePath,
-    temporary: true,
-    filePath: path.basename(filePath)
+    location: token.temporary
   }]
   token.parsed = true
 }
