+++ lib/util/archive.js
@@ -1,11 +1,10 @@
+const path = require('path')
 const execa = require('execa')
 
-const {createTempDirectory} = require('./tmpdir')
+async function unarchive(location) {
+  const temp = path.join(path.dirname(location), '_unarchived')
 
-async function unarchive(filePath) {
-  const temp = await createTempDirectory()
-
-  await execa('unar', ['-no-directory', '-output-directory', temp, filePath])
+  await execa('unar', ['-no-directory', '-output-directory', temp, location])
 
   return temp
 }
